requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
drf-spectacular==0.27.1
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.3.1